Finished debounced processor. Also, Maven!

+++ src/main/java/googleinterview/debounce/Clock.java
@@ -1,5 +1,5 @@
 package googleinterview.debounce;
 
 public interface Clock {
-  long getTime();
+  int getTime();
 }
